+++ tests/test_trait.py
@@ -16,16 +16,86 @@
 
 from __future__ import annotations
 
+from operator import attrgetter
+from string import ascii_lowercase, digits
+
 import pytest
 
-from extra_platforms import ALL_TRAITS, UNKNOWN
+import extra_platforms
+from extra_platforms import ALL_GROUP_IDS, ALL_IDS, ALL_TRAIT_IDS, ALL_TRAITS, UNKNOWN
 from extra_platforms.trait import Trait
 
+all_traits_params = pytest.mark.parametrize(
+    "trait", list(ALL_TRAITS | UNKNOWN), ids=attrgetter("id")
+)
 
-def test_all_traits_have_detection_function():
-    for trait in ALL_TRAITS | {UNKNOWN}:
-        assert trait.detection_func_id
-        assert trait.current in (True, False)
+
+@all_traits_params
+def test_all_traits_generated_constants(trait):
+    assert trait
+
+    assert trait.id
+    assert trait.id.isascii()
+    assert trait.id[0] in ascii_lowercase
+    assert trait.id[-1] in ascii_lowercase + digits
+    assert set(trait.id).issubset(ascii_lowercase + digits + "_")
+    assert trait.id.islower()
+    # Traits are not allowed to start with all_ or any_, which is reserved for groups.
+    assert not trait.id.startswith(("all_", "any_"))
+    assert trait.id not in ALL_GROUP_IDS
+    if trait.id.startswith("unknown"):
+        assert trait in UNKNOWN
+        assert trait.id not in ALL_TRAIT_IDS
+        assert trait.id not in ALL_IDS
+        assert trait.id == f"unknown_{trait.__class__.__name__.lower()}"
+        assert (
+            trait.name.lower()
+            == "unknown " + trait.__class__.__name__.replace("_", " ").lower()
+        )
+        assert trait.icon == "❓"
+    else:
+        assert trait.id in ALL_TRAIT_IDS
+        assert trait.id in ALL_IDS
+        assert trait not in UNKNOWN
+
+    # Name.
+    assert trait.name
+    assert trait.name.isascii()
+    assert trait.name.isprintable()
+
+    # Icon.
+    assert trait.icon
+    assert 2 >= len(trait.icon) >= 1
+
+    # URL.
+    assert trait.url
+    assert trait.url.startswith("https://")
+
+    # Symbol ID.
+    assert trait.symbol_id
+    assert trait.symbol_id == trait.id.upper()
+    assert hasattr(extra_platforms, trait.symbol_id)
+
+    # Detection function.
+    assert trait.detection_func_id
+    assert hasattr(extra_platforms, trait.detection_func_id)
+    assert trait.current in (True, False)
+
+    # Info.
+    assert trait.info()
+    for k, v in trait.info().items():
+        assert set(k).issubset(ascii_lowercase + "_")
+        if v is not None:
+            assert isinstance(v, (str, bool, dict))
+            if isinstance(v, str):
+                assert v
+            elif isinstance(v, dict):
+                assert v
+                for k1, v1 in v.items():
+                    assert set(k1).issubset(ascii_lowercase + "_")
+                    if v1 is not None:
+                        assert v1
+    assert trait.info()["id"] == trait.id
 
 
 def test_detection_function_missing(caplog):